Update login try pages -> components
